Commit 8
This is a quite heavy commit.
this commit is aimed to make our code's  user login/registration modular and easy to be read and make changes to , in this we are adding multiple layers in our to our app so that Firebase layer code is not exposed to the top UI layer code.
The aim is to have below layering ->
# User Interface
# Service
# Providers
# Auth Provider abstract class , Auth User class and Auth Exceptions

>In the auth exceptions file we have created our won exceptions instead of using directly the firebase exceptions/ any other provider. (might seem confusing but as we are not doing anything with the exception data returned from firebase exception/any other provider , it is totally fine to replace it with our exceptions)
>In the auth user file we have created a factory constructor which takes a firebase /any other provider user and uses it's email verified property and creates another User defined by us as an AuthUser.
>In auth provider file we have created an abstract class with methods we want our providers to fulfill
>In the FirebaseAuth provider file we have created actual logic in which we communicate with the firebase and implement all the methods from the abstract class
>Now the service file will use an auth provider and expose it's functions to the top layer
Hence instead of directly using firebase we are using our own service which may have some additional logic and manipulation for the data provided by the providers like firebase, google, apple etc.