Commit 10
This commit includes the testing suite build for the application. In this we created a mock auth provider which we have control on and then we implement all the methods in it to be able to fake how firebase works. This was done to perform unit test on our auth service.
After making a mock auth provider we created a group of different tests and checked the expected values using the expect function inside the test function.

For detailed code and it's working please see the code file as the implementation is simple and easy to understand