Commit 22
Objective of this commit is to learn about BloC and BloC Library , It is a library / set of components made by Very Good Ventures.
This library helps to separate the business logic from the main UI of the app this has its own benefits like:
- More readable code
- Iteration and testing become easy
- Code becomes more modular and making new changes and adding features get easy
- Reduces the information flow between UI an logic layer and thus makes the application more secure and robust

Below are the important concepts of BloC :
- It has events and state , the events are the inputs to the BloC and states are the output of the BloC.
- Events can be actions taken by user with the UI.
- Based on the events the BloC generates the output as state that changes the state of the application.
- To use a BloC we are supposed to have states of same type and events of same type , ie. they inherit the same parent event or state.
- Then we have listener and builder , basically the listener contains code that needs to be executed every time the state of the application changes and builder is then called next which returns the widget to the Scaffold of the application.
- We also need to create our BloC using the BloC class from the library. Inside it we call the constructor of the super class and pass the default state of the application , or the initial state of the application.
- Then we can use the "on" function which contains the code to be run on the event type defined while declaring the on function , inside it we use emit function to set/change the state of the application.

In our app we comment out the already existing code and  wrote a new home page to learn about BloC (and for this we created a Counter app , which upon entering the text in the text field update the counter value).
More can be understood by glancing at the code and written notes