Commit 25
This is a very long commit and is focused on the concept of Bloc and adopting it.
In this commit we continued our process of moving towards the Bloc code style this helps to isolate the UI and the business logic while at the same time providing a bridge of communication between them.
In this commit we not only expanded the bloc usage but also create new functionalities in terms of states and evet and also modified and removed certain states and events.

One the that bloc has made very effort less is navigation in the application , earlier navigator was used to move from one view to other but now since we are using a bloc builder in the home page , every time the state changes the builder runs and checks the state and returns the corresponding view.

Below are the bloc states and events :
1.We removed Loading state and instead made provision of loading dialog for it.
2.We created a new Uninitialized state for when the application starts.
3.creatd a isLoading parameter in the LoggedOut state.
4.Also now as the LoggedOut state is representing 3 states we used a package called equatable to be able to make the LoggedOut class equatable by including the line List<Objects?>get props => [exception,isLoading] , which means that if any new instance is created and differ in these properties then it will be considered as a different state.
5.Also created a Registering state which is emitted with and without exception based whether the registering was successful or not.
6.We created new email verification event along with register and should register events.
7. We created a new file for the loading dialog , this dialog will be shown whenever the application made contact with firebase, also this dialog returned a function which can be called to close this dialog. Though it may seem a good idea it must change as the dialog should itself acknowledge the completion of firebase action and dismiss itself.
For this we created a function showLoadingDialog of return type CloseDialog which is a typedef for a function of return type void.
inside it we created an alert dialog and then showed this dialog on screen using the showDialog function with barrierDismissible property set to false so that the user can not dismiss this loading screen. After this , we return a inline function which pops the dialog using the navigator.
So when this function will be called by the user , say after the login process is complete, it will pop the dialog from the navigation stack.