Commit 26
This commit builds up on the previous commit and the main aim is to discover and migrate to new way of displaying the loading screen.

In the previous commit we used an Alert Dialog to construct our loading screen/pop-up and a Show Dialog to render the dialog on the screen. To understand the problem with this method we first have to understand working of stack in  flutter application. Any Material page (view/scaffold/container) we display is sent to the Navigation stack, and upon moving from one page to other or showing dialogs , those components (ie. the pages or the dialogs) are added to the stack , so if we want to go back to the previous page which was being rendered before pushing/displaying the current page/dialog we need to pop the current page from the stack.
The problem in using a loading screen via Alert Dialog or Show Dialog is that it will be pushed into the stack and to remove it we will have to use pop() function , this function is not aware the  component we are popping. So if we showed loading screen for multiple things by pushing it into the stack  or , pushed another component while the current loading screen is present at top of the stack , upon using the pop() function the new added component will be removed instead of the loading screen we wanted to be removed.
In an application there are many instances where we need to show such loading screens while also presenting some other components therefore Navigation Stack is not a good place to push our loading dialogs.

The solution to this is to use another type Stack called Overlay stack in which we push the Overlays which are independent of the Navigation Stack. The entries in this Overlay stack can be seen as widgets similar to those in Navigation Stack. The benefits of using Overlays is that while dismissing the loading screen the intended loading screen will be removed and Navigation Stack will stay unaffected.

Task done->
# Made a new folder named helper which has another folder named loading and inside it we have two files to create and manage our loading screen , they are , "loading_screen_controller.dart" and "loading_screen.dart".

# Inside the first file we created a class which handles two functions one to close the loading screen and another to update the loading screen. This is done by first defining the Alias for the functions , the close function is a bool function and the update function is a bool function which takes a string as parameter (this string will be used to update the loading screen text).Then we created an immutable class called LoadingScreenController which has two member variables called close and update of the type CloseLoadingScreen and UpdateLoadingScreen respectively ,(the alias we made for the functions). And then we created a constructor for the class which set the value for these members. That is upon calling the class we need to pass the function which will be assigned to the close and update members.
Note:-It is a good trick to use a class and create member variables to handle functions when there is a need to have multiple functions to operate on similar stuff (in this case to work on a loading screen)

# Next in the loading screen file we made a class named LoadingScreen , the aim is to have functionality of creating loading screen using this class and also have functions that can update or close the loading screen.

# We first made the class a singleton as we only want a single instance of it to be present in the application.

# Then we created a function named "showOverlays" with return type of LoadingScreenController, it accepts a build context and a string as parameters.

# Inside this function a String stream controller is created , to which we push our text , to be displayed on the loading screen , any text in future is added to this stream.

# Then we use the Overlay.of(context) funciton to get the Overlay State of our application which will help us to add Overlay Entries into the Overlay Stack.

# Then we used findRenderObject function to get the render box of the application. Render box can be thought of as an object which represents the space available to render the widgets on, later extracted the size of the render box using its size property

# Then created a OverlayEntry object which takes a required builder function and needs a widget to be returned.

# This function returns a Material widget which takes the whole page, in which we have a container at center with constraints for dimensions and some decoration. For the child widget of the container we have SingleChildScrollView() which lets us create a single scrollable widget in which we have a column containing SizedBox, CircularProgresIndicator ,SizedBox and then a Stream Builder which builds a Text widget based on the text stream. This Text widget forms the label for the loading screen. If no text is available in the stream then empty container is shown.

# Then we add this overlay entry into the overlay stack using the overlay state we grabbed for the current context. This is done using the insert function .

# Then we return our LoadingScreenController with functions for close and update parameters. in close function we first closed the text stream controller and then removed the overlay and returned true, and for the update function we simply added the new text in the stream controller and returned true.

Note:- We will use this showOverlay function internally.

# Created a null LoadingScreenController handle with null as default value.

# Then inside the LoadingScreen class we created two new functions named show and hide. The show function checks whether controller is present or not, ie. whether the loading screen is being displayed or not, if it is being displayed then a new text is sent to stream controller using update function and if if  loading screen is not shown then a new overlay is created.

# In the hide function screen is simply closed using the close function.

# Next in the Auth State file we added two final variables named isLoading of type bool and text of type String? . The isLoading variable will act as a global variable an looking at this our application will show or hide the loading screen. We created a constructor for the abstract class AuthState to initialize the final variables , the default value for the text is also set.

# Made changes accordingly in all other states.

# In the logged out state we also added a text variable and passed it to the  super constructor.

# In the auth bloc file we made necessary updates in the states.

# Next we removed the previous loading screen implementation, as it used Alert dialog and now we have moved to using overlays.

# Using Overlay function helps us to have the overlay logic isolated from rest of the files unlike the previous implementation in which we need to create handle and logic and each file.

# It might seem convenient to have the loading screen overlay logic inside the builder function but it is a bad practice because overlays are side-effects and therefore must be used inside the listener.

# So in order to use both the builder and listener inside the main function we rather use BlocConsumer which provides with both the builder and listener.

# The builder code remains the same and inside the listener we check whether isLoading property f the state is true or false , if true we use the loading screen show function to show the loading screen and if the isLoading property is set to false then we hide the loading screen.

Our application is almost complete , we are just required to make some necessary quality improvements like providing a reset password button in case the user forgets the password , etc..