Commit 28
In this commit we changed our application name and set a new icon for the application and also added the option for splash screen , there was no code as such written , as most of it was done via GUI of VSCode and Figma where we designed our splash screen.
(To know the details please refer hard notes or the video).